Do not generate entire taskgraph.  Instead just generate the slice that a particular processor cares about.  This creates a nice chicken/egg problem that we will have to worry about in a dynamic load-balancing world, but for now we will just focus on getting to that point...

+++ Packages/Uintah/CCA/Components/Schedulers/MemoryLog.h
@@ -11,7 +11,7 @@ namespace Uintah {
 		 const std::string& label, const std::string& name,
 		 const std::string& type, const Patch* patch,
 		 int material, const std::string& elems,
-		 int size, void* ptr, int dwid=-1);
+		 unsigned long size, void* ptr, int dwid=-1);
 }
 
 #endif
